test: split the context module by concern
