+++ tests/integration/platform/data_daemon/behavioural_correctness/test_signal_cleanup.py
@@ -41,12 +41,14 @@ from tests.integration.platform.data_daemon.shared.test_case.build_test_case imp
     DataDaemonTestCase,
     case_ids,
 )
-from tests.integration.platform.data_daemon.shared.test_case.build_test_case_context import (  # noqa: E501
-    build_context_specs,
-    run_case_contexts,
-)
 from tests.integration.platform.data_daemon.shared.test_case.constants import (
     STOP_METHOD_SIGKILL,
+)
+from tests.integration.platform.data_daemon.shared.test_case.context_spec import (
+    build_context_specs,
+)
+from tests.integration.platform.data_daemon.shared.test_case.context_worker import (
+    run_case_contexts,
 )
 from tests.integration.platform.data_daemon.shared.test_infrastructure import (
     cloud_resource_deleter,
