+++ tests/integration/platform/data_daemon/shared/test_case/context_spec.py
@@ -1,0 +1,343 @@
+"""What one context is asked to do, and what it reports having done."""
+
+from __future__ import annotations
+
+import random
+import uuid
+from dataclasses import dataclass, field
+
+from neuracore.core.streaming.recording_state_manager import RecordingStateManager
+from tests.integration.platform.data_daemon.shared.test_case.boundaries import (
+    ObservedFrameCodes,
+    TraceClassification,
+)
+from tests.integration.platform.data_daemon.shared.test_case.build_test_case import (
+    DataDaemonTestCase,
+    case_id,
+)
+from tests.integration.platform.data_daemon.shared.test_case.constants import (
+    DURATION_MODE_VARIABLE,
+    DURATION_VARIABLE_MAX_FACTOR,
+    DURATION_VARIABLE_MIN_FACTOR,
+    MODE_STAGGERED,
+    STOP_RECORDING_NO_WAIT_SLA_S,
+    STOP_RECORDING_OVERHEAD_PER_SEC,
+    STOP_RECORDING_UPLOAD_SLA_PER_JOINT_SAMPLE_S,
+    STOP_RECORDING_UPLOAD_SLA_PER_VIDEO_PIXEL_S,
+    DepthMode,
+    random_phase_jitter_window,
+)
+
+CONTEXT_DURATION_RANDOM = random.Random(0)
+
+# Stream discriminators feeding ``stream_phase_seed`` so a recording's joint and
+# video streams draw independent phase offsets.
+JOINT_STREAM = 0
+VIDEO_STREAM = 1
+
+
+@dataclass(frozen=True, slots=True)
+class RecordingExpectedTimestamps:
+    """Expected timestamps per trace for one recording, keyed by semantic trace name.
+
+    Produced during the recording loop (once the recording key is known) and
+    consumed by :func:`~disk_helpers.assert_disk_recording_properties`
+    to verify on-disk trace.json files match the timestamps that were logged.
+    These are the *same lists* the producer emitted, so the assertion is exact
+    equality regardless of ``random_phase``.
+
+    Attributes:
+        by_trace: Maps semantic trace key (e.g. ``"JOINT_POSITIONS"``,
+            ``"camera_0"``) to the :class:`~boundaries.TraceClassification` the
+            recording's boundary rule returned for it.
+    """
+
+    by_trace: dict[str, TraceClassification]
+
+
+@dataclass(frozen=True, slots=True)
+class ContextExpectedTimestamps:
+    """Expected timestamps for all recordings produced by one context worker.
+
+    Attributes:
+        by_recording: Maps the on-disk recording directory name to its
+            :class:`RecordingExpectedTimestamps`. The directory name is the
+            integer ``recording_index`` as a string.
+    """
+
+    by_recording: dict[str, RecordingExpectedTimestamps]
+
+
+@dataclass(frozen=True, slots=True)
+class ContextCaseSpec:
+    duration_sec: int
+    joint_count: int
+    producer_channels: str
+    video_count: int
+    image_width: int | None
+    image_height: int | None
+    joint_fps: int
+    video_fps: int
+    wait: bool
+    random_phase: bool
+    video_detail: str
+    producer_pacing: str
+    depth_count: int = 0
+    depth_mode: DepthMode = "float32"
+
+    @property
+    def stop_recording_sla_s(self) -> float:
+        """Seconds allowed for the ``nc.stop_recording`` call.
+
+        ``wait=False`` is fire-and-forget — the call never blocks on the
+        upload pipeline — so it gets a flat constant. ``wait=True`` blocks
+        until every trace has uploaded, so its budget is the sum of the
+        joint-data and video-data upload costs: total joint samples
+        (``duration_sec * joint_count * joint_fps``) and total video pixels
+        across both RGB and depth cameras (``duration_sec * video_fps *
+        (video_count + depth_count) * image_width * image_height``), each
+        times an observed per-unit upload cost. Depth cameras reuse the RGB
+        per-pixel upload constant as a first approximation — both are
+        video-family traces that always keep a lossless archive, so their
+        upload cost is comparable order-of-magnitude, though not necessarily
+        identical. The budget is floored at the duration-based overhead so
+        short or low-volume recordings keep a sane minimum.
+        """
+        if not self.wait:
+            return STOP_RECORDING_NO_WAIT_SLA_S
+        duration_floor = self.duration_sec * STOP_RECORDING_OVERHEAD_PER_SEC
+        joint_budget = (
+            self.duration_sec
+            * self.joint_count
+            * self.joint_fps
+            * STOP_RECORDING_UPLOAD_SLA_PER_JOINT_SAMPLE_S
+        )
+        video_budget = 0.0
+        camera_count = self.video_count + self.depth_count
+        if camera_count and self.image_width and self.image_height:
+            video_budget = (
+                self.duration_sec
+                * self.video_fps
+                * camera_count
+                * self.image_width
+                * self.image_height
+                * STOP_RECORDING_UPLOAD_SLA_PER_VIDEO_PIXEL_S
+            )
+        return max(duration_floor, joint_budget + video_budget)
+
+
+@dataclass(frozen=True, slots=True)
+class ContextResult:
+    """Per-context result from a completed recording workload.
+
+    Produced by :func:`context_worker` and consumed by assertion helpers
+    and verification functions throughout the test suite.
+
+    A recording is addressed by:
+
+    - ``recording_ids`` — the cloud ``recording_id`` (TEXT) for each recording.
+      These are what cloud verification (``verify_cloud_results``) matches
+      against the dataset's ``recording.id``. The daemon mints them
+      asynchronously, so an entry may be an empty string until the test
+      resolves it (via ``resolve_cloud_recording_ids``) once online.
+
+    - ``recording_indexes`` — the daemon-assigned local INTEGER
+      ``recording_index`` for each recording, resolved from the source DB.
+      These are the on-disk directory names and the daemon-DB join key.
+    - ``source`` is the ``(robot_id, robot_instance)`` identity used to correlate
+      a worker's recordings to daemon-minted ``recording_index`` values without
+      relying on the local handle.
+    """
+
+    dataset_name: str
+    recording_ids: list[str]
+    robot_name: str
+    joint_names: list[str]
+    camera_names: list[str]
+    joint_frame_count: int
+    video_frame_count: int
+    joint_fps: int
+    video_fps: int
+    duration_sec: int
+    timestamp_start_s: float
+    timestamp_end_s: float
+    marker_names: list[str]
+    has_video: bool
+    context_index: int
+    wall_started_at: float | None
+    wall_stopped_at: float
+    random_phase: bool
+    expected_timestamps: ContextExpectedTimestamps = field(
+        default_factory=lambda: ContextExpectedTimestamps(by_recording={})
+    )
+    timer_stats: dict[str, dict[str, float]] = field(default_factory=dict)
+    recording_indexes: list[int] = field(default_factory=list)
+    source: tuple[str, int] = ("", 0)
+    depth_camera_names: list[str] = field(default_factory=list)
+    depth_frame_count: int = 0
+    depth_mode: DepthMode = "float32"
+    has_depth: bool = False
+    observed_frame_codes: dict[str, ObservedFrameCodes] = field(default_factory=dict)
+    """Painted camera frame codes per recording, keyed by ``recording_index``."""
+    expected_video_stop_timestamp_by_recording: dict[str, float] = field(
+        default_factory=dict
+    )
+    """Nominal capture-clock upper bound of each recording's video."""
+
+
+@dataclass(frozen=True, slots=True)
+class ContextSpec:
+    case: ContextCaseSpec
+    context_index: int
+    robot_name: str
+    dataset_name: str
+    recordings_per_context: int
+    expected_joint_frames: int
+    expected_video_frames: int
+    timestamp_start_s: float
+    timestamp_end_s: float
+    assert_deadline: bool = False
+
+
+def build_context_specs(
+    case: DataDaemonTestCase,
+    dataset_name: str | None = None,
+    assert_deadline: bool = False,
+) -> list[ContextSpec]:
+    """Build per-context worker specs for a matrix case.
+
+    ``assert_deadline`` arms the ``Timer`` deadline assertions; what rate the
+    producer offers frames at is the case's own ``producer_pacing``.
+    """
+    specs: list[ContextSpec] = []
+    timestamp_stagger_s = case.duration_sec / 2.0
+    base_recordings_per_context = case.recording_count // case.parallel_contexts
+    recording_remainder = case.recording_count % case.parallel_contexts
+    shared_dataset_name = (
+        dataset_name or f"testing_dataset_{case_id(case)}_{uuid.uuid4().hex[:6]}"
+    )
+
+    for context_index in range(case.parallel_contexts):
+        timestamp_start_s = 0.0
+        if context_index > 0 and case.mode == MODE_STAGGERED:
+            timestamp_start_s = float(timestamp_stagger_s * context_index)
+
+        if case.context_duration_mode == DURATION_MODE_VARIABLE:
+            context_duration_sec = max(
+                1,
+                min(
+                    int(
+                        case.duration_sec
+                        * CONTEXT_DURATION_RANDOM.uniform(
+                            DURATION_VARIABLE_MIN_FACTOR, DURATION_VARIABLE_MAX_FACTOR
+                        )
+                    ),
+                    RecordingStateManager.MAX_RECORDING_DURATION_S,
+                ),
+            )
+        else:
+            context_duration_sec = case.duration_sec
+
+        recordings_for_context = base_recordings_per_context + (
+            1 if context_index < recording_remainder else 0
+        )
+
+        specs.append(
+            ContextSpec(
+                case=ContextCaseSpec(
+                    duration_sec=context_duration_sec,
+                    joint_count=case.joint_count,
+                    producer_channels=case.producer_channels,
+                    video_count=case.video_count,
+                    image_width=case.image_width,
+                    image_height=case.image_height,
+                    joint_fps=case.joint_fps,
+                    video_fps=case.video_fps,
+                    wait=case.wait,
+                    random_phase=case.random_phase,
+                    video_detail=case.video_detail,
+                    producer_pacing=case.producer_pacing,
+                    depth_count=case.depth_count,
+                    depth_mode=case.depth_mode,
+                ),
+                context_index=context_index,
+                robot_name=f"matrix_robot_{uuid.uuid4().hex[:10]}",
+                dataset_name=shared_dataset_name,
+                recordings_per_context=recordings_for_context,
+                expected_joint_frames=case.joint_fps * context_duration_sec,
+                expected_video_frames=case.video_fps * context_duration_sec,
+                timestamp_start_s=timestamp_start_s,
+                timestamp_end_s=(
+                    timestamp_start_s + context_duration_sec * recordings_for_context
+                ),
+                assert_deadline=assert_deadline,
+            )
+        )
+    return specs
+
+
+def stream_phase_seed(context_index: int, recording_ordinal: int, stream: int) -> int:
+    """Return a stable RNG seed for one recording's stream.
+
+    Distinct per (context, recording, stream) so parallel contexts and the joint
+    and video streams within a recording all draw independent phase offsets,
+    while a rerun of the same case reproduces them exactly.
+    """
+    return (context_index * 1_000 + recording_ordinal) * 2 + stream
+
+
+def precompute_timestamps(
+    timestamp_start_s: float,
+    frame_count: int,
+    fps: int,
+    random_phase: bool = False,
+    seed: int = 0,
+) -> list[float]:
+    """Return the complete synthetic timestamp sequence for one stream.
+
+    Frames sit on an exact ``timestamp_start_s + frame_index / fps`` grid.  With
+    *random_phase* each frame is additionally offset by a pseudo-random amount
+    within :func:`random_phase_jitter_window`, so the daemon sees non-uniformly
+    spaced timestamps.  The offsets are drawn from a *seed*-derived RNG, which is
+    what lets the caller hand the very same list to both the producer and the
+    on-disk expectation.
+    """
+    if not random_phase:
+        return [
+            timestamp_start_s + frame_index / fps for frame_index in range(frame_count)
+        ]
+
+    rng = random.Random(seed)
+    window = random_phase_jitter_window(fps)
+    return [
+        timestamp_start_s + frame_index / fps + rng.uniform(-window, window)
+        for frame_index in range(frame_count)
+    ]
+
+
+def recording_timestamps(
+    spec: ContextSpec, recording_index: int
+) -> tuple[list[float], list[float]]:
+    """Return the ``(joint, video)`` timestamp sequences for one recording.
+
+    Shares nothing with the schedule producers emit from, so a bounded producer
+    is checked against what it was supposed to log rather than against itself.
+    """
+    case = spec.case
+    start_s = spec.timestamp_start_s + recording_index * case.duration_sec
+    return (
+        precompute_timestamps(
+            start_s,
+            spec.expected_joint_frames,
+            case.joint_fps,
+            random_phase=case.random_phase,
+            seed=stream_phase_seed(spec.context_index, recording_index, JOINT_STREAM),
+        ),
+        precompute_timestamps(
+            start_s,
+            spec.expected_video_frames,
+            case.video_fps,
+            random_phase=case.random_phase,
+            seed=stream_phase_seed(spec.context_index, recording_index, VIDEO_STREAM),
+        ),
+    )
